add basic crawler worker
This adds a couple of classes, the main one being the Crawler worker.
This sidekiq worker is intended to use its adapters to connect to social
networks and fetch posts that include the #wearemagma hashtag. After
fetching them, it stores them and separate the new from the previously
stored ones. Finally, it sends the new posts to the hubot so it can
notifiy the Slack channel.