remove DB model
This removes the DB model and all references to it from the crawler app.

I think the crawler should have a proper db instead of relying on redis to store
posts. I don't need the crawler to store posts for now, so I'm removing this
model to later add a proper database and a proper ActiveRecord-like model.